bot/requirements.txt
aiodns==3.0.0
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
Babel==2.9.1
backoff==2.2.1
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cohere==4.11.2
exceptiongroup==1.1.3
filelock==3.12.2
frozenlist==1.3.3
h11==0.14.0
hiredis==2.2.3
httpcore==1.0.2
httpx==0.25.1
idna==3.4
importlib-metadata==6.7.0
Jinja2==3.1.2
magic-filter==1.0.9
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
pycares==4.3.0
pycparser==2.21
pydantic==2.0
pydantic_core==2.0.1
pydub==0.25.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
regex==2023.6.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sympy==1.12
typing_extensions==4.6.3
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.3
yarl==1.9.2
zipp==3.15.0
